fix: settling
