+++ src/components/Bets/BetTablePageSelector.tsx
@@ -1,4 +1,4 @@
-import { decrementPage, incrementPage } from "../../../utils/bets";
+import { decrementPage, incrementPage } from "../../utils/bets";
 
 type Props = {
   page: number;
@@ -6,7 +6,7 @@ type Props = {
   setPage: (page: number) => void;
 };
 
-const PageSelector: React.FC<Props> = ({ page, maxPages, setPage }) => (
+const BetTablePageSelector: React.FC<Props> = ({ page, maxPages, setPage }) => (
   <div className="w-full flex justify-end col-span-2 font-semibold mt-2 select-none">
     <div className="bg-white flex p-1 border-gray-300 border rounded-md">
       <div className="mr-4">
@@ -52,4 +52,4 @@ const PageSelector: React.FC<Props> = ({ page, maxPages, setPage }) => (
   </div>
 );
 
-export default PageSelector;
+export default BetTablePageSelector;
