+++ src/components/Bets/index.ts
@@ -1,0 +1,4 @@
+export { default as BetRows } from "./BetRows";
+export { default as BetTable } from "./BetTable";
+export { default as BetTablePageSelector } from "./BetTablePageSelector";
+export { default as BetRow } from "./BetRow";
