+++ src/components/Modals/index.ts
@@ -1,0 +1,2 @@
+export { default as PlaceBetModal } from "./PlaceBetModal";
+export { default as SettleBetModal } from "./SettleBetModal";
