+++ src/components/index.ts
@@ -1,5 +1,5 @@
 import QuickForm from "./QuickForm/QuickForm_View";
-import Loader from "./Loader/Loader_View";
+import Loader from "./Loader";
 import Button from "./Button/Button_View";
 import Input from "./Input/Input_View";
 import QuickFormInput from "./QuickFormInput/QuickFormInput_Logic";
